requirements.txt
aiogram==2.24
aiohttp==3.8.3
aiosignal==1.3.1
aiosqlite==0.18.0
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
asyncselenium==0.0.2
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
outcome==1.2.0
pycparser==2.21
PySocks==1.7.1
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
selenium==3.141.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.14
wsproto==1.2.0
yarl==1.8.2
